Fixes and updates to the test suite
The test suite now runs the diff after it complete the test, this allows
us to know the results before the diff is output.  If the test passed
then all results files, except for the leaves_results, for use in the
bottom up test, will be deleted.  Otherwise they are kept for review
The top down test has a built in delay so as to allow the writer to
finish writing to the file before running tests on a potentially
incomplete file.
There is a bug where certain files aren't being deleted by the program,
I do not know why.