Cleaned ReprDataModel of unused methods
Tests now run at about half speed, performance hit isn't a problem, just
should be noted.
ReprDataModel was cleaned of unused or redundant methods, it is now a
third the size it used to be and readability has been improved as method
calls are now not several layers deep to write to file.
Fixed bug where leaves weren't being saved.