Move xmlns declarations to root element
The XML namespace declarations in Fiwalk have been at the wrong point
in the XML tree.  Being where they were, only one or two elements would
get any namespace assigned to them.

Applying this patch, the output of Fiwalk is now valid DFXML according
to the (draft) XML Schema at:
<https://github.com/dfxml-working-group/dfxml_schema/blob/v1.1.0rfc0/dfxml.xsd>

After some discussion with the DFXML Working Group, it will be
appropriate to bump the /dfxml/@version attribute to "1.1.0" to indicate
support of an official DFXML Schema version; however, the current
schema version is a Request for Comments.  Input on the schema is
welcome.

Signed-off-by: Alex Nelson <a.nelson@prometheuscomputing.com>