fiwalk: Update rusage reporting
The <runstats> element was an early version of a resource usage report.
Later versions, particularly the one employed by Hashdeep, simply use an
<rusage> element with no renaming of its fields, and an extra
<clocktime> element.

This change makes Fiwalk output a consistent element with Hashdeep and
the DFXML C++ library.  It adds precision and programmatic access to the
elapsed time record.

This and the prior patch also cause Fiwalk's output to validate against
the DFXML schema version here:
<https://github.com/ajnelson/dfxml_schema/tree/archive/for_hashdeep_v1>

Adding getrusage shouldn't cause any problems elsewhere in the codebase;
the HAVE_GETRUSAGE symbol doesn't appear anywhere.

(This change does not use the DFXML library code included in the same
directory, because that is a less-than-straightforward inclusion. A
future patch should make use of dfxml.{h,cpp} as libraries.)

Signed-off-by: Alex Nelson <a.nelson@prometheuscomputing.com>