Remove escaping of strings from prepared statement
I haven't tested this, but the change but it seems simple enough.
It shouldn't be necessary to escape the string when using a PreparedStatement.  PreparedStatments handle escaping for you.  The escaping causes any single quotes to be saved as two single quotes in the database, an issue that I've observed.