bindings/java/build-unix.xml: do not check for the static library presence
When the build system is configured with '--enable-java
--disable-static', it will not generate the 'libtsk_jni.a' file, even
though the shared library (.so file) is still created. This
configuration results in ant failing with "JNI native library not
built". However, the check for the static library is not meaningful
since only the shared library is of interest for the JNI.

Rewrite the logic so that it checks for the presence of either the
dylib or the so file.