Refactor UsnJrnl logic
The refactoring enables the support for Journal records of version 3
and 4.

The declarations have been all moved to tsk_ntfs.h as they are
internal for the moment.

A generic TSK_USN_RECORD structure has been introduced. It will rely
on a union to contain the version specific data.

The version specific data is contained in TSK_USN_V*_DATA structures.
As the V2 and V3 structures contain the same information, they've been
grouped into a single TSK_USN_V3_DATA to signify the highes supported
version.

The TSK_USN_RECORD_HEADER is now merged into the TSK_USN_RECORD and
can be used by the User to determine the union type (v2, v3, v4).

This simplifies the callback logic as it now receives a
TSK_USN_RECORD structure only.

Signed-off-by: Matteo Cafasso <noxdafox@gmail.com>