Fix for CT-2917 - fail to get non-ascii registry key names
Non-ascii registry key names seem to not be Null terminated. I created three non-ascii keys all of which had no Null termination. I verified manually using a hex editor on the registry hive. This led to conversion issues when calling from_bytes() within getUFT16String(). Solution is to append a UFT16 null char to the end of any string we try to do a conversion on that is not null terminated.