API updates newBlackboardArtifact
newBlackboardArtifact takes a collection of attributes and a boolean allowDuplicates flag
transaction may be null. if it is null a new transaction will be created and used
removed Blackboard#addBlackboardAttributes method. This function does not update the
Artifact object cache and can potentially cause issues if the caller is not aware of this.
Added BlackboardArtifact#addAttributes method that does update its cache.
The downside of this change is that the caller needs to load an Artifact into memory if
it does not have it. This API is however safer as the caller will have the opportunity to
validate or duplicate check the attribute etc.