Addressing SQL performance issues
The index defined for tsk_host_addresses are not used by SQLITE (per the Explain Query Plan output)
when the query has a LOWER defined on it.
Also, the in query is not useful as we determine the type when the insert is done.
The updates made here engages the index to perform the query.