Upgrade fiwalk to redesignate byte_runs facet for data
Currently, Fiwalk generates byte_runs to describe where file contents
reside on the disk.

    <byte_runs>
      <byte_run img_offset="...".../>
      ...
    </byte_runs>

There are reasons to make byte_runs to represent other parts of a file.
To support when other byte_runs are made, we are going to support the
current byte_runs to include a new attribute. This is an additive,
backwards compatible change.

    <byte_runs facet='data'>
      <byte_run img_offset="...".../>
      ...
    </byte_runs>

References:
* In the Python DFXML objects.py bindings, see the properties
  FileObject.data_brs and ByteRuns.facet.
* This supports implementation of an API described in Appendix A of:
  https://doi.org/10.1016/j.diin.2019.06.004

Licensing:
Portions of this patch contributed by NIST are governed by the NIST
Software Licensing Statement:

NIST-developed software is provided by NIST as a public service. You may
use, copy, and distribute copies of the software in any medium, provided
that you keep intact this entire notice. You may improve, modify, and
create derivative works of the software or any portion of the software,
and you may copy and distribute such modifications or works. Modified
works should carry a notice stating that you changed the software and
should note the date and nature of any such change. Please explicitly
acknowledge the National Institute of Standards and Technology as the
source of the software.

NIST-developed software is expressly provided "AS IS." NIST MAKES NO
WARRANTY OF ANY KIND, EXPRESS, IMPLIED, IN FACT, OR ARISING BY OPERATION
OF LAW, INCLUDING, WITHOUT LIMITATION, THE IMPLIED WARRANTY OF
MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, NON-INFRINGEMENT, AND
DATA ACCURACY. NIST NEITHER REPRESENTS NOR WARRANTS THAT THE OPERATION
OF THE SOFTWARE WILL BE UNINTERRUPTED OR ERROR-FREE, OR THAT ANY DEFECTS
WILL BE CORRECTED. NIST DOES NOT WARRANT OR MAKE ANY REPRESENTATIONS
REGARDING THE USE OF THE SOFTWARE OR THE RESULTS THEREOF, INCLUDING BUT
NOT LIMITED TO THE CORRECTNESS, ACCURACY, RELIABILITY, OR USEFULNESS OF
THE SOFTWARE.

You are solely responsible for determining the appropriateness of using
and distributing the software and you assume all risks associated with
its use, including but not limited to the risks and costs of program
errors, compliance with applicable laws, damage to or loss of data,
programs or equipment, and the unavailability or interruption of
operation. This software is not intended to be used in any situation
where a failure could cause risk of injury or damage to property. The
software developed by NIST employees is not subject to copyright
protection within the United States.

Co-authored-by: Alex Nelson <alexander.nelson@nist.gov>
Cc: Sheldon Douglas <sheldoug19@gmail.com>
Signed-off-by: Sheldon Douglas <sheldon.douglas@nist.gov>