Modify ntfs_dinode_lookup calls to export start_of_inode
Two of the six ntfs_dinode_lookup calls were selected from observing
access to a TSK_FS_META struct: ntfs_inode_lookup, ntfs_inode_walk.
This patch starts the production of the chain that enables fiwalk to
generate byte_runs for a file's inode. This is an additive, backwards
compatible change.

Modifications were tested against three disk images:
* DFTT Test 3
* M57 Patents, "Jo", November 12, start
* M57 Patents, "Jo", November 20, OldComputer

The test method was seeing if the inode byte_run would give an offset
that had the ASCII bytes "FILE".

The test succeeded DFFT Test 3, which is an image of a file system
with no containing partition system. That is, offset 0 in the image
is the beginning of the file system.

The test does not yet succeed for the "Jo" images, which have
partition systems.

A follow-on patch will correct the issues.

References:
* This supports implementation of an API described in Appendix A of:
  https://doi.org/10.1016/j.diin.2019.06.004
* http://dftt.sourceforge.net/test3/index.html
* https://digitalcorpora.org/corpora/scenarios/m57-patents-scenario/

Licensing:
Portions of this patch contributed by NIST are governed by the NIST
Software Licensing Statement:

NIST-developed software is provided by NIST as a public service. You may
use, copy, and distribute copies of the software in any medium, provided
that you keep intact this entire notice. You may improve, modify, and
create derivative works of the software or any portion of the software,
and you may copy and distribute such modifications or works. Modified
works should carry a notice stating that you changed the software and
should note the date and nature of any such change. Please explicitly
acknowledge the National Institute of Standards and Technology as the
source of the software.

NIST-developed software is expressly provided "AS IS." NIST MAKES NO
WARRANTY OF ANY KIND, EXPRESS, IMPLIED, IN FACT, OR ARISING BY OPERATION
OF LAW, INCLUDING, WITHOUT LIMITATION, THE IMPLIED WARRANTY OF
MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, NON-INFRINGEMENT, AND
DATA ACCURACY. NIST NEITHER REPRESENTS NOR WARRANTS THAT THE OPERATION
OF THE SOFTWARE WILL BE UNINTERRUPTED OR ERROR-FREE, OR THAT ANY DEFECTS
WILL BE CORRECTED. NIST DOES NOT WARRANT OR MAKE ANY REPRESENTATIONS
REGARDING THE USE OF THE SOFTWARE OR THE RESULTS THEREOF, INCLUDING BUT
NOT LIMITED TO THE CORRECTNESS, ACCURACY, RELIABILITY, OR USEFULNESS OF
THE SOFTWARE.

You are solely responsible for determining the appropriateness of using
and distributing the software and you assume all risks associated with
its use, including but not limited to the risks and costs of program
errors, compliance with applicable laws, damage to or loss of data,
programs or equipment, and the unavailability or interruption of
operation. This software is not intended to be used in any situation
where a failure could cause risk of injury or damage to property. The
software developed by NIST employees is not subject to copyright
protection within the United States.

Co-authored-by: Alex Nelson <alexander.nelson@nist.gov>
Cc: Sheldon Douglas <sheldoug19@gmail.com>
Signed-off-by: Sheldon Douglas <sheldon.douglas@nist.gov>