Don't overlink executables.
Dynamically linked executables were being linked with libtsk _and_ libtsk's
dependencies, which is unnecessary and was obscuring that libtsk wasn't
being linked with its dependencies. The image format tests were designed
to check that libtsk is able to use the dependent libraries which read
image formats. Linking the tools used in the image format tests with
the dependent libraries prevents the tests from failing when libtsk wasn't
linked to those dependent libraries.